pyproject.toml
[project]
name = "trifecta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "fastapi[standard]>=0.115.6",
    "httpx>=0.28.1",
    "pyyaml>=6.0.2",
]
